+++ copasi/compareExpressions/CNormalLogical.cpp
@@ -1,4 +1,14 @@
-// Copyright (C) 2010 - 2013 by Pedro Mendes, Virginia Tech Intellectual
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
+// University of Virginia, University of Heidelberg, and University
+// of Connecticut School of Medicine.
+// All rights reserved.
+
+// Copyright (C) 2017 - 2018 by Pedro Mendes, Virginia Tech Intellectual
+// Properties, Inc., University of Heidelberg, and University of
+// of Connecticut School of Medicine.
+// All rights reserved.
+
+// Copyright (C) 2010 - 2016 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc., University of Heidelberg, and The University
 // of Manchester.
 // All rights reserved.
@@ -898,7 +908,7 @@ bool CNormalLogical::generateCanonicalDNF(ItemSetOfSets& tmpAndSets) const
               // create a new row for the truth table
               // the bits in i can be mapped to the truth values
               std::bitset<16> bitSet((unsigned long long)i);
-              unsigned int j, jMax = itemVector.size();
+              unsigned int j, jMax = (unsigned int)itemVector.size();
 
               for (j = 0; j < jMax; ++j)
                 {
